test (security): Refactor security tests and remove deprecated files
This commit refactors the security tests by organizing them into dedicated directories for the engine, key store, services, and KMS.
It introduces new test files that cover the functionality of the security engine, key store, and KMS, ensuring better modularity
and maintainability of the test suite.
It also removes deprecated test files that are no longer relevant.
It covers all the cases of the security engine, key store, and KMS, ensuring that the security features are thoroughly tested.
Also it use `localstack` for testing aws services instead of real AWS services, which allows for faster and more reliable tests
without the need for network access.