test (storage & security): Enhance integration tests and add security
This commit enhances the integration tests for storage by split the tests of the local storage from the AWS storage tests.
It replace test with real aws services with `localstack` to ensure that the tests are not dependent on external services.
This allows for more reliable and faster tests, as they can be run in isolation without needing to connect to AWS.
It also adds a new integration test for security features, ensuring that the security mechanisms are functioning as expected.
Through all the options kms, local key store , and google key store.