Fix (integrity path): Add the full path to the integrity files
This commit updates the integrity path in the checksum and HMAC integrity files to ensure that the correct paths are
used during integrity checks. This change is necessary for the proper functioning of the integrity verification process.
It also remove passing path for integrity file in HMAC integrity check, as the path is now hardcoded
from the current processing directory.